test: reorganize and consolidate relay_setup tests
Created tests/test_relay_setup.py and tests/test_relay_schema.py.
Moved CLI tests to tests/test_cli.py.
Removed redundant test files.
