+++ tests/test_relay_setup.py
@@ -1,4 +1,4 @@
-"""Tests for relay_schema, relay_setup, and credential_state modules."""
+"""Tests for relay_setup module."""
 
 from __future__ import annotations
 
@@ -15,9 +15,7 @@ from better_code_review_graph.relay_setup import (
     ensure_config,
 )
 
-# Test URL used when exercising the create_session path. Matches the
-# MCP_RELAY_URL env var set by the module-level autouse fixture below
-# (no production default per mode-matrix 2.5).
+# Test URL used when exercising the create_session path.
 TEST_RELAY_URL = "https://relay.example.com"
 
 
@@ -26,99 +24,18 @@ def _default_relay_url_env(monkeypatch):
     """Set MCP_RELAY_URL for every test in this module.
 
     Per mode-matrix 2.5, remote-relay mode requires explicit MCP_RELAY_URL
-    (no DEFAULT_RELAY_URL fallback). Tests that need to verify the "missing
-    MCP_RELAY_URL" behavior can delenv it inside the test body.
+    (no DEFAULT_RELAY_URL fallback).
     """
     monkeypatch.setenv("MCP_RELAY_URL", TEST_RELAY_URL)
 
 
-# ---------------------------------------------------------------------------
-# relay_schema
-# ---------------------------------------------------------------------------
-
-
-class TestRelaySchema:
-    def test_server_name(self):
-        assert RELAY_SCHEMA["server"] == "better-code-review-graph"
-
-    def test_display_name(self):
-        assert RELAY_SCHEMA["displayName"] == "Code Review Graph"
-
-    def test_has_fields(self):
-        fields = RELAY_SCHEMA["fields"]
-        assert len(fields) == 4
-
-    def test_field_keys(self):
-        keys = [f["key"] for f in RELAY_SCHEMA["fields"]]
-        assert keys == [
-            "JINA_AI_API_KEY",
-            "GEMINI_API_KEY",
-            "OPENAI_API_KEY",
-            "COHERE_API_KEY",
-        ]
-
-    def test_all_fields_optional(self):
-        for field in RELAY_SCHEMA["fields"]:
-            assert field["required"] is False
-
-    def test_schema_is_valid_typed_dict(self):
-        assert "server" in RELAY_SCHEMA
-        assert "displayName" in RELAY_SCHEMA
-        assert "fields" in RELAY_SCHEMA
-
-
-# ---------------------------------------------------------------------------
-# relay_setup
-# ---------------------------------------------------------------------------
-
-
-class TestRelaySetupConstants:
-    def test_server_name(self):
-        assert SERVER_NAME == "better-code-review-graph"
-
-    def test_cloud_keys(self):
-        assert "JINA_AI_API_KEY" in CLOUD_KEYS
-        assert "GEMINI_API_KEY" in CLOUD_KEYS
-        assert "OPENAI_API_KEY" in CLOUD_KEYS
-        assert "COHERE_API_KEY" in CLOUD_KEYS
-
-    # DEFAULT_RELAY_URL removed per mode-matrix 2.5 (no centralized subdomain
-    # for default-local servers). See test_missing_relay_url_raises below for
-    # the enforcement test.
-
-
-class TestEnsureConfigEnvVar:
-    async def test_returns_none_when_env_set(self, monkeypatch):
-        monkeypatch.setenv("GEMINI_API_KEY", "AIzaFakeKey123")
+class TestEnsureConfig:
+    async def test_env_vars_priority(self, monkeypatch):
+        monkeypatch.setenv("GEMINI_API_KEY", "existing")
         result = await ensure_config()
-        assert result is None
+        assert result is None  # env vars take priority, relay skipped
 
-    async def test_any_cloud_key_takes_priority(self, monkeypatch):
-        for key in CLOUD_KEYS:
-            monkeypatch.delenv(key, raising=False)
-        monkeypatch.setenv("JINA_AI_API_KEY", "jina_test")
-        result = await ensure_config()
-        assert result is None
-
-    async def test_empty_env_var_not_accepted(self, monkeypatch):
-        for key in CLOUD_KEYS:
-            monkeypatch.delenv(key, raising=False)
-        monkeypatch.setenv("GEMINI_API_KEY", "")
-
-        with patch(
-            "better_code_review_graph.relay_setup.PerPluginStore"
-        ) as mock_store_cls:
-            mock_store_cls.return_value.load.return_value = None
-            with patch(
-                "mcp_core.relay.client.create_session",
-                side_effect=Exception("no relay"),
-            ):
-                result = await ensure_config()
-                assert result is None
-
-
-class TestEnsureConfigFile:
-    async def test_returns_config_from_file(self, monkeypatch):
+    async def test_config_loaded_from_store(self, monkeypatch):
         for key in CLOUD_KEYS:
             monkeypatch.delenv(key, raising=False)
 
@@ -126,46 +43,14 @@ class TestEnsureConfigFile:
             "better_code_review_graph.relay_setup.PerPluginStore"
         ) as mock_store_cls:
             mock_store_cls.return_value.load.return_value = {
-                "GEMINI_API_KEY": "from-file"
+                "GEMINI_API_KEY": "from-store"
             }
             result = await ensure_config()
-            assert result is not None
-            assert result["GEMINI_API_KEY"] == "from-file"
-            mock_store_cls.assert_called_once_with(SERVER_NAME)
+            assert result == {"GEMINI_API_KEY": "from-store"}
+            assert os.environ.get("GEMINI_API_KEY") == "from-store"
+        monkeypatch.delenv("GEMINI_API_KEY", raising=False)
 
-    async def test_injects_config_into_env(self, monkeypatch):
-        for key in CLOUD_KEYS:
-            monkeypatch.delenv(key, raising=False)
-
-        with patch(
-            "better_code_review_graph.relay_setup.PerPluginStore"
-        ) as mock_store_cls:
-            mock_store_cls.return_value.load.return_value = {
-                "GEMINI_API_KEY": "injected-key"
-            }
-            await ensure_config()
-            assert os.environ.get("GEMINI_API_KEY") == "injected-key"
-            monkeypatch.delenv("GEMINI_API_KEY", raising=False)
-
-    async def test_returns_none_when_no_cloud_keys_in_file(self, monkeypatch):
-        for key in CLOUD_KEYS:
-            monkeypatch.delenv(key, raising=False)
-
-        with patch(
-            "better_code_review_graph.relay_setup.PerPluginStore"
-        ) as mock_store_cls:
-            mock_store_cls.return_value.load.return_value = {"UNKNOWN_KEY": "value"}
-            with patch(
-                "mcp_core.relay.client.create_session",
-                side_effect=Exception("no relay"),
-            ):
-                result = await ensure_config()
-                assert result is None
-
-
-class TestEnsureConfigRelay:
-    async def test_missing_relay_url_raises(self, monkeypatch):
-        """Remote-relay without MCP_RELAY_URL must raise per matrix 2.5."""
+    async def test_relay_url_required(self, monkeypatch):
         for key in CLOUD_KEYS:
             monkeypatch.delenv(key, raising=False)
         monkeypatch.delenv("MCP_RELAY_URL", raising=False)
@@ -189,6 +74,7 @@ class TestEnsureConfigRelay:
             mock_store_cls.return_value.save = MagicMock()
             mock_session = MagicMock()
             mock_session.relay_url = "https://relay.example.com/setup#k=abc"
+            mock_session.session_id = "test-sess"
 
             with (
                 patch(
@@ -201,8 +87,11 @@ class TestEnsureConfigRelay:
                     new_callable=AsyncMock,
                     return_value={"JINA_AI_API_KEY": "from-relay"},
                 ) as mock_poll,
-                patch("httpx.AsyncClient") as _mock_http,
+                patch("httpx.AsyncClient") as mock_http_cls,
             ):
+                mock_http = mock_http_cls.return_value.__aenter__.return_value
+                mock_http.post = AsyncMock()
+
                 result = await ensure_config()
                 assert result is not None
                 assert result["JINA_AI_API_KEY"] == "from-relay"
@@ -213,6 +102,7 @@ class TestEnsureConfigRelay:
                 mock_store_cls.return_value.save.assert_called_once_with(
                     {"JINA_AI_API_KEY": "from-relay"}
                 )
+        monkeypatch.delenv("JINA_AI_API_KEY", raising=False)
 
     async def test_relay_server_unreachable(self, monkeypatch):
         for key in CLOUD_KEYS:
@@ -281,11 +171,6 @@ class TestEnsureConfigRelay:
             ):
                 result = await ensure_config()
                 assert result is None
-
-
-# ---------------------------------------------------------------------------
-# CLI integration -- tests serve_main calls resolve_credential_state
-# ---------------------------------------------------------------------------
 
 
 class TestApplyConfig:
@@ -399,32 +284,82 @@ class TestEnsureConfigRelayNotifyFailure:
                 assert result is None
 
 
-class TestCLIIntegration:
-    def test_main_runs_stdio_directly(self):
-        """serve_main(stdio) invokes FastMCP stdio directly (no bridge layer)."""
-        from better_code_review_graph import server as server_module
+# Tests from test_relay_setup_coverage_fix.py
 
-        with (
-            patch.object(server_module.mcp, "run") as mock_run,
-            patch.dict(os.environ, {"MCP_TRANSPORT": "stdio"}),
+
+@pytest.mark.asyncio
+async def test_read_config_exception(monkeypatch):
+    for key in CLOUD_KEYS:
+        monkeypatch.delenv(key, raising=False)
+
+    # PerPluginStore constructor raises Exception -- should fall through to relay
+    with patch(
+        "better_code_review_graph.relay_setup.PerPluginStore",
+        side_effect=Exception("Disk error"),
+    ):
+        # To avoid going into relay setup, we also mock create_session to fail
+        with patch(
+            "mcp_core.relay.client.create_session",
+            side_effect=Exception("no relay"),
         ):
-            from better_code_review_graph.cli import main
+            result = await ensure_config()
+            assert result is None
 
-            main()
-            mock_run.assert_called_once_with(transport="stdio")
 
-    def test_main_continues_on_relay_error(self):
-        with (
-            patch(
-                "better_code_review_graph.credential_state.resolve_credential_state",
-                side_effect=Exception("relay broken"),
-            ),
-            patch("better_code_review_graph.server.mcp"),
-            patch.dict(os.environ, {"MCP_TRANSPORT": "stdio"}),
+@pytest.mark.asyncio
+async def test_httpx_post_exception(monkeypatch):
+    for key in CLOUD_KEYS:
+        monkeypatch.delenv(key, raising=False)
+
+    mock_session = MagicMock()
+    mock_session.relay_url = "https://relay.example.com/setup"
+    mock_session.session_id = "test-session"
+
+    mock_client = MagicMock()
+    mock_client.post = AsyncMock(side_effect=Exception("Network error"))
+    mock_client.__aenter__ = AsyncMock(return_value=mock_client)
+    mock_client.__aexit__ = AsyncMock(return_value=False)
+
+    with patch("better_code_review_graph.relay_setup.PerPluginStore") as mock_store_cls:
+        mock_store_cls.return_value.load.return_value = None
+        with patch(
+            "mcp_core.relay.client.create_session",
+            new_callable=AsyncMock,
+            return_value=mock_session,
         ):
-            from better_code_review_graph.cli import main
+            with patch(
+                "mcp_core.relay.client.poll_for_result",
+                new_callable=AsyncMock,
+                return_value={"GEMINI_API_KEY": "new-key"},
+            ):
+                with patch("httpx.AsyncClient", return_value=mock_client):
+                    result = await ensure_config()
+                    assert result is not None
+                    assert result["GEMINI_API_KEY"] == "new-key"
+                    assert os.environ.get("GEMINI_API_KEY") == "new-key"
+    monkeypatch.delenv("GEMINI_API_KEY", raising=False)
 
-            try:
-                main()
-            except Exception:
-                pass
+
+@pytest.mark.asyncio
+async def test_poll_for_result_runtime_error_unexpected(monkeypatch):
+    for key in CLOUD_KEYS:
+        monkeypatch.delenv(key, raising=False)
+
+    with patch("better_code_review_graph.relay_setup.PerPluginStore") as mock_store_cls:
+        mock_store_cls.return_value.load.return_value = None
+        mock_session = MagicMock()
+        mock_session.relay_url = "https://relay.example.com/setup"
+
+        with patch(
+            "mcp_core.relay.client.create_session",
+            new_callable=AsyncMock,
+            return_value=mock_session,
+        ):
+            # RuntimeError with unexpected message -- should return None
+            with patch(
+                "mcp_core.relay.client.poll_for_result",
+                new_callable=AsyncMock,
+                side_effect=RuntimeError("something went wrong"),
+            ):
+                result = await ensure_config()
+                assert result is None
